foton_core::block_entity::trialspawner

Trait TrialSpawnerStateAccessor

pub trait TrialSpawnerStateAccessor {
    // Required methods
    fn trial_spawner_state(&self) -> TrialSpawnerState;
    fn set_trial_spawner_state(
        &self,
        world: &Arc<World>,
        state: TrialSpawnerState,
    );
    fn mark_trial_spawner_updated(&self);
}

The block state a trial spawner reads and writes.

Vanilla parity: TrialSpawner.StateAccessor, implemented by the block entity because the state lives on the block, not on the spawner.

Required Methods

fn trial_spawner_state(&self) -> TrialSpawnerState

Vanilla parity: StateAccessor.getState.

fn set_trial_spawner_state(&self, world: &Arc<World>, state: TrialSpawnerState)

Vanilla parity: StateAccessor.setState.

fn mark_trial_spawner_updated(&self)

Vanilla parity: StateAccessor.markUpdated.

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".