foton_core::entity::entities::mobs

Module passive

Passive entity implementations.

Re-exports

pub use equine::CamelEntity;

pub use equine::CamelHuskEntity;

pub use equine::DonkeyEntity;

pub use equine::HorseEntity;

pub use equine::HorseMarkings;

pub use equine::HorseVariant;

pub use equine::LlamaEntity;

pub use equine::MuleEntity;

pub use equine::SkeletonHorseEntity;

pub use equine::TraderLlamaEntity;

pub use equine::ZombieHorseEntity;

Modules

allay 🔒

Those mobs are passive creatures that run away when attacked by a player. Allay entity.

armadillo 🔒

Armadillo entity.

axolotl 🔒

Axolotl entity.

bee 🔒

Bee entity.

cat 🔒

Cat entity.

chicken 🔒

Chicken entity.

cow 🔒

Vanilla Cow entity with variant + sound-variant parity.

equine

The horse family.

fox 🔒

Fox entity.

frog 🔒

Frog entity.

goat 🔒

Goat entity.

happy_ghast 🔒

Happy ghast entity.

mooshroom 🔒

Vanilla Mooshroom entity (MushroomCow) with variant, stew feeding, and shearing parity.

nautilus 🔒

The nautilus family.

ocelot 🔒

Ocelot entity.

panda 🔒

Panda entity.

parrot 🔒

Parrot entity.

pig 🔒

Pig entity implementation.

polar_bear 🔒

Polar bear entity.

rabbit 🔒

Rabbit entity.

sheep 🔒

Vanilla Sheep entity with wool color, shearing, and breeding parity.

sniffer 🔒

Sniffer entity.

strider 🔒

Strider entity.

turtle 🔒

Turtle entity.

Structs

AllayEntity

An allay.

ArmadilloEntity

An armadillo.

AxolotlEntity

An axolotl.

BeeEntity

A bee.

CatEntity

Vanilla cat entity.

ChickenEntity

A chicken.

CowEntity

Vanilla cow entity with synced variant and sound-variant state.

FoxEntity

Vanilla fox entity.

FrogEntity

A frog.

GoatEntity

A goat.

HappyGhastEntity

A happy ghast.

MushroomCowEntity

Vanilla mooshroom entity: a Shearable AbstractCow subclass with a synced RED/BROWN variant and an unsynced suspicious-stew-effect payload.

NautilusEntity

A nautilus.

OcelotEntity

Vanilla ocelot entity.

PandaEntity

A panda.

ParrotEntity

Vanilla parrot entity.

PigEntity

Vanilla pig entity.

PolarBearEntity

A polar bear.

RabbitEntity

A rabbit.

SheepEntity

Vanilla sheep entity.

SnifferEntity

A sniffer.

StriderEntity

A strider.

TurtleEntity

A turtle.

ZombieNautilusEntity

A zombie nautilus.

Enums

FoxVariant

The two coats a fox comes in.

MushroomCowVariant

Vanilla parity: MushroomCow.Variant.

PandaGene

One of the seven things a panda can be.

ParrotVariant

The five colors a parrot comes in.

RabbitVariant

Vanilla Rabbit.Variant.

Functions

hatch_sniffer_from_egg

Hatches a baby sniffer out of a SnifferEggBlock.

spawn_allay

Spawns an allay for a test or a command.