foton_core::entity::entities::objects

Module projectiles

Projectile entity implementations.

Modules

arrow 🔒

Arrow entity.

breeze_wind_charge 🔒

A breeze’s wind charge.

dragon_fireball 🔒

Dragon fireball.

ender_pearl 🔒

Thrown ender pearl projectile entity (ThrownEnderpearl).

evoker_fangs 🔒

Evoker fangs entity.

experience_bottle 🔒

Thrown bottle o’ enchanting.

eye_of_ender 🔒

Eye of ender entity.

firework_rocket 🔒

Firework rocket projectile entity (FireworkRocketEntity).

fishing_hook 🔒

The fishing bobber (FishingHook).

large_fireball 🔒

Ghast fireball.

lingering_potion 🔒

Thrown lingering potion.

llama_spit 🔒

A llama’s spit.

shulker_bullet 🔒

Shulker bullet.

small_fireball 🔒

Blaze fireball.

snowball 🔒

Thrown snowball.

spectral_arrow 🔒

Spectral arrow entity.

splash_potion 🔒

Thrown splash potion.

thrown_egg 🔒

Thrown egg.

thrown_trident 🔒

Thrown trident entity.

wind_charge 🔒

Wind charge.

wind_charge_common 🔒

What the two wind charges share.

wither_skull 🔒

Wither skull.

Structs

ArrowEntity

A flying arrow.

BreezeWindChargeEntity

A wind charge a breeze fired.

DragonFireballEntity

A dragon’s fireball.

EnderPearlEntity

A thrown ender pearl.

EvokerFangsEntity

A pair of evoker fangs.

ExperienceBottleEntity

A thrown bottle o’ enchanting.

EyeOfEnderEntity

A thrown eye of ender.

FireworkRocketEntity

A launched firework rocket.

FishingHookEntity

A cast fishing bobber.

LargeFireballEntity

A ghast’s fireball.

LingeringPotionEntity

A thrown lingering potion.

LlamaSpitEntity

A llama’s spit.

ShulkerBulletEntity

A shulker’s homing bullet.

SmallFireballEntity

A blaze’s fireball.

SnowballEntity

A thrown snowball.

SpectralArrowEntity

A flying spectral arrow.

SplashPotionEntity

A thrown splash potion.

ThrownEggEntity

A thrown egg.

ThrownTridentEntity

A trident in flight.

WindChargeEntity

A thrown wind charge.

WitherSkullEntity

A wither’s skull.

Enums

ArrowPickup

Who is allowed to pick an arrow back up.

TridentPickup

Who, if anyone, may pick this trident back up.