foton_core::event

Module entity_lifecycle

Entities arriving in the world and leaving their vehicles. Events about entities arriving in the world and leaving their vehicles.

Structs

EntitiesLoadEvent

Entities saved in a chunk were loaded back into the world.

EntitiesUnloadEvent

The entities of a chunk are about to be unloaded with it.

EntityDismountEvent

An entity is getting off what it rides.

EntityPlaceEvent

A player placed an entity with an item: a boat, a minecart, an armor stand, an end crystal.