foton_core::world::game_event::vibrations::user

Trait VibrationUser

pub trait VibrationUser: Send + Sync {
    // Required methods
    fn listener_radius(&self) -> i32;
    fn position_source(&self) -> VibrationPositionSource;
    fn can_receive_vibration(
        &self,
        world: &Arc<World>,
        pos: BlockPos,
        event: GameEventRef,
        context: &GameEventContext<'_>,
    ) -> bool;
    fn on_receive_vibration(
        &self,
        world: &Arc<World>,
        pos: BlockPos,
        event: GameEventRef,
        source_entity: Option<&dyn Entity>,
        projectile_owner: Option<&dyn Entity>,
        receiving_distance: f32,
    );

    // Provided methods
    fn listenable_events(&self) -> Identifier { ... }
    fn can_trigger_avoid_vibration(&self) -> bool { ... }
    fn requires_adjacent_chunks_to_be_ticking(&self) -> bool { ... }
    fn calculate_travel_time_in_ticks(
        &self,
        distance_to_destination: f32,
    ) -> i32 { ... }
    fn is_valid_vibration(
        &self,
        event: GameEventRef,
        context: &GameEventContext<'_>,
    ) -> bool { ... }
    fn on_data_changed(&self) { ... }
}

Vanilla VibrationSystem.User.

One implementation per thing that hears vibrations: the two sculk sensors, the shrieker, the warden and the allay.

Required Methods

fn listener_radius(&self) -> i32

Vanilla VibrationSystem.User.getListenerRadius.

fn position_source(&self) -> VibrationPositionSource

Vanilla VibrationSystem.User.getPositionSource.

fn can_receive_vibration( &self, world: &Arc<World>, pos: BlockPos, event: GameEventRef, context: &GameEventContext<'_>, ) -> bool

Vanilla VibrationSystem.User.canReceiveVibration.

fn on_receive_vibration( &self, world: &Arc<World>, pos: BlockPos, event: GameEventRef, source_entity: Option<&dyn Entity>, projectile_owner: Option<&dyn Entity>, receiving_distance: f32, )

Vanilla VibrationSystem.User.onReceiveVibration.

Provided Methods

fn listenable_events(&self) -> Identifier

Vanilla VibrationSystem.User.getListenableEvents.

fn can_trigger_avoid_vibration(&self) -> bool

Vanilla VibrationSystem.User.canTriggerAvoidVibration.

fn requires_adjacent_chunks_to_be_ticking(&self) -> bool

Vanilla VibrationSystem.User.requiresAdjacentChunksToBeTicking.

fn calculate_travel_time_in_ticks(&self, distance_to_destination: f32) -> i32

Vanilla VibrationSystem.User.calculateTravelTimeInTicks.

This is the delay that makes a sensor feel like it is hearing something rather than touching it: one tick per block between the source and the listener.

fn is_valid_vibration( &self, event: GameEventRef, context: &GameEventContext<'_>, ) -> bool

Vanilla VibrationSystem.User.isValidVibration.

Not implemented: the AVOID_VIBRATION advancement trigger a sneaking player earns here, because Foton has no advancement criteria system.

fn on_data_changed(&self)

Vanilla VibrationSystem.User.onDataChanged.

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".